+++ app/src/main/java/com/example/schoolproject/test/DataBaseHelper.java
@@ -88,7 +88,7 @@ public class DataBaseHelper extends SQLiteOpenHelper {
         if (cursor != null && cursor.moveToFirst()){
             do {
                 DataPost data1 = new DataPost();
-                data1.setPostId(cursor.getInt(cursor.getColumnIndex("postId")));
+                data1.setPostId(cursor.getLong(cursor.getColumnIndex("postId")));
                 data1.setBoardType(cursor.getString(cursor.getColumnIndex("boardType")));
                 data1.setTitle(cursor.getString(cursor.getColumnIndex("title")));
                 data1.setContent(cursor.getString(cursor.getColumnIndex("content")));
